+++ JOSS/paper.bib
@@ -78,7 +78,7 @@ volume = {29},
 number = {11},
 pages = {1859-1865},
 keywords = {molecular dynamics, Poisson-Boltzmann, electrostatic potential, membrane, visualization, explicit solvent, implicit solvent, MarvinSpace},
-doi = {https://doi.org/10.1002/jcc.20945},
+doi = {10.1002/jcc.20945},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.20945},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.20945},
 abstract = {Abstract CHARMM is an academic research program used widely for macromolecular mechanics and dynamics with versatile analysis and manipulation tools of atomic coordinates and dynamics trajectories. CHARMM-GUI, http://www.charmm-gui.org, has been developed to provide a web-based graphical user interface to generate various input files and molecular systems to facilitate and standardize the usage of common and advanced simulation techniques in CHARMM. The web environment provides an ideal platform to build and validate a molecular model system in an interactive fashion such that, if a problem is found through visual inspection, one can go back to the previous setup and regenerate the whole system again. In this article, we describe the currently available functional modules of CHARMM-GUI Input Generator that form a basis for the advanced simulation techniques. Future directions of the CHARMM-GUI development project are also discussed briefly together with other features in the CHARMM-GUI website, such as Archive and Movie Gallery. © 2008 Wiley Periodicals, Inc. J Comput Chem, 2008},
@@ -89,7 +89,7 @@ year = {2008}
 author = {UNESCO},
 title = {UNESCO Recommendation on Open Science},
 howpublished = {Electronic Paper},
-doi = {https://doi.org/10.54677/MNMH8546},
+doi = {10.54677/MNMH8546},
 url = {https://unesdoc.unesco.org/ark:/48223/pf0000379949.locale=en},
 year = {2021}
 }
