### 2024-01-12 #6219 @moritzobenauer
Hello everybody,

in the following is a list of potential reviewers I find suitable for the project:
[Xujian Wang](https://reviewers.joss.theoj.org/reviewers/3552)
[Benjamin Ries](https://reviewers.joss.theoj.org/reviewers/3554)[richardjgowers](https://reviewers.joss.theoj.org/reviewers/107)
[Yuxuan Zhuang](https://reviewers.joss.theoj.org/reviewers/2144)
[Matt Thompson](https://reviewers.joss.theoj.org/reviewers/1138)

If I can be of any further help, please let me know. Looking forward to the reviewing process. Best, Moritz
