### 2024-01-16 #6219 @Kevin-Mattheus-Moerman
@moritzobenauer thanks for this submission. I am the AEiC on this track and here to help with initial steps. I have just flagged this work for a scope review. This is because I need some help from the domain specialists in the editorial board to judge if this submission is in scope for JOSS.  This scope review should take about two weeks.
