### 2024-01-23 #6219 @Kevin-Mattheus-Moerman
@moritzobenauer I am happy to inform you that the editorial board has determined that this submission may be in scope for JOSS. I have removed the query-scope label and will now be looking to assign a handling editor.
