+++ JOSS/paper.bib
@@ -113,6 +113,7 @@ year = {2021}
   pages={1--17},
   year={2012},
   publisher={BioMed Central}
+  doi={10.1186/1758-2946-4-17}
 }
 
 @book{python3,
