+++ JOSS/paper.bib
@@ -112,7 +112,7 @@ year = {2021}
   number={1},
   pages={1--17},
   year={2012},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
   doi={10.1186/1758-2946-4-17}
 }
 
