### 2024-01-28 #6293 @editorialbot
[REVIEW]: Project RACCOON: Automated construction of PDB files for polymers and polymer peptide conjugates
**Submitting author:** @moritzobenauer (<a href="http://orcid.org/0009-0003-8140-9907">Moritz Obenauer</a>)
**Repository:** https://github.com/moritzobenauer/ProjectRaccoon
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
