### 2024-02-04 #6293 @taoliu032
@moritzobenauer Hi Moritz, firstly thanks for putting together such a package! Before getting into the paper and codebase, I want to clarify what seems like a small issue in the list of authors. In the article proof, one author is tagged with *, indicating co-authorship. I guess at least one more author should also be labeled with * ?
