### 2024-02-04 #6293 @moritzobenauer
> @moritzobenauer Hi Moritz, firstly thanks for putting together such a package! Before getting into the paper and codebase, I want to clarify what seems like a small issue in the list of authors. In the article proof, one author is tagged with *, indicating co-authorship. I guess at least one more author should also be labeled with * ?

Hey taoliu032,

You're absolutely right. I thought my corresponding authorship would imply an equal contribution. I updated the paper. Thank you for the hint and sorry for the oversight.
