### 2024-02-21 moritzobenauer/ProjectRaccoon#10 @lorenzo-rovigatti
Can't run the suggested example
Hi there! I'm currently working on reviewing this software for JOSS (see the [review thread](https://github.com/openjournals/joss-reviews/issues/6293#issuecomment-1956725433)).
I have tried installing the lib with pip (`pip install project-raccoon==1.0.1`), and the installation seems to complete successfully. However, if I try to run the example (`project_raccoon -s examples/seq_FHFHFXG_PEO_GXFHFHF.txt`) I get the following error:

```
Traceback (most recent call last):
  File "/home/lorenzo/software/anaconda3/bin/project_raccoon", line 5, in <module>
    from project_raccoon.__main__ import main
  File "/home/lorenzo/software/anaconda3/lib/python3.11/site-packages/project_raccoon/__init__.py", line 1, in <module>
    from .src.data import Monomer, Monomers, Sequence, Atom
ImportError: cannot import name 'Monomer' from 'project_raccoon.src.data' (unknown location)
```

Here is the output of `pip freeze`: [pip_freeze.txt](https://github.com/moritzobenauer/ProjectRaccoon/files/14365178/pip_freeze.txt)

If I try to use the suggested fallback install through conda I get this output:

```
$ conda env create -f environment.yml
Retrieving notices: ...working... done
Solving environment: failed

ResolvePackageNotFound:
  - ca-certificates==2023.11.17=h8857fd0_0
  - sqlite==3.44.2=h7461747_0
  - tk==8.6.13=h1abcd95_1
  - openssl==3.2.0=hd75f5a5_1
  - readline==8.2=h9e318b2_1
  - expat==2.5.0=hf0c8a7f_1
  - libexpat==2.5.0=hf0c8a7f_1
  - bzip2==1.0.8=h10d778d_5
  - libsqlite==3.44.2=h92b6c6a_0
  - ncurses==6.4=h93d8f39_2
  - python==3.12.0=hd58486a_0
  - libffi==3.4.2=h0d85af4_5
  - zlib==1.2.13=h8a1eda9_5
  - xz==5.4.5=h6c40b1e_0
  - libzlib==1.2.13=h8a1eda9_5
```

For reference, I'm using anaconda v 23.7.4. Now I'm lost: what should I do? :-)
