### 2024-02-22 moritzobenauer/ProjectRaccoon#10 @kainszs
Good morning,

I was able to reproduce the error messages, and there seem to be several sources of error.

1. The conda environment was created with osx64 software. I will take care of this later.
2. I need to look into the pip installation in more detail, so far it is unclear to me why it is not working properly.

To test the package independently of pip, I suggest the following steps:
   - clone the repository: git clone https://github.com/moritzobenauer/ProjectRaccoon.git
   - create an environment with Python >=3.11: conda create -n raccoon_env python=3.11
   - install the app from the directory: pip install -e .

I hope this helps and I apologize for the inconvenience.
