### 2024-02-22 moritzobenauer/ProjectRaccoon#10 @kainszs
The problem is probably due to the fact that there was no __init__.py in the source order and therefore the whole module was not installed. I have changed this, but would only create the new version in consultation with you after the package has been fully tested.
