### 2024-02-22 moritzobenauer/ProjectRaccoon#10 @lorenzo-rovigatti
Thank you! Using a fresh conda env and installing from the repo works. Note that there is a typo in the README (`conda activate racoon` should read `conda activate raccoon`).
About the new version, it probably makes sense to release it at the end of the review process. I'll close this one for now!
