### 2024-02-23 moritzobenauer/ProjectRaccoon#12 @lorenzo-rovigatti
JOSS review: documentation & testing
Hi there, here is another issue linked to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6293#issuecomment-1956725433)

The documentation as it is now is fully contained in the README file and makes it possible to easily install the software and use it on the example seq.txt file. However, there are some details missing that should be added to help future users:

1. how to add monomers? There is some info in the `monomers.dat` file, but I reckon it would be better to but all the documentation in the same place
2. how to test the code?
3. what to do with the other files and folders that are in the repo? What's in the `gmx` folder? What's in the `examples` folder beyond the example seq.txt?
