### 2024-02-23 #6293 @lorenzo-rovigatti
First of all, thanks to @moritzobenauer (and the other authors) for this nice piece of software. I have added three issues to the repo, with one of them which has been already closed.

While the authors work on the the other issues, I have a couple of comments on some specific items that can be more easily discussed here:

1. According to the commit history, the code has been authored by @moritzobenauer and @kainszs . Did the other authors contribute to the code in some other ways?
2. Community guidelines (see the checklist above) are missing.
