### 2024-02-24 moritzobenauer/ProjectRaccoon#11 @kainszs
Good morning,

These are valid points. However, it may be a few days before I can take care of it as I am moving this weekend. I see from your GIthub that you also have expertise in Python. Can you recommend a library for the parser, or is it recommended writing it yourself? Are self-written exceptions sufficient, or do you know a more elegant way to catch the errors and output error messages? How would you approach this?
