### 2024-02-24 #6293 @kainszs
Hi,
we have noticed that with increasingly larger sequences, we get performance issues when generating the polymer.

We are testing the module against the example sequence. We have noticed that the actions sometimes abort due to a runtime error, i.e. that the tests exceed the 6h limit. This is the case with almost every fourth test.

You can run the tests via

```bash
python -m unittest
```

I attribute this to the fact that the generation of the sequence by the self-avoiding-random walk is in a continuous loop because no displacement vector is found that fulfills the minimum distance condition.

The polymer is generated in ``project_raccoon/src/functions/standard.py`` using the functions '``generate_file()`` and ``get_semi_random_walk_shift()``. We store all coordinates of the polymer in a matrix and generate a vector $\vec{k}$ so that the coordinates of the new monomer buildingblock added with $\vec{k}$ exceed at least the threshold as distance. Do you have any idea how we can fix this without increasing the threshold? We would be grateful for any ideas.

![aborted_action](https://github.com/openjournals/joss-reviews/assets/100122494/c83a54c7-aaf6-47d0-ae7d-363bf7608653)

 @lorenzo-rovigatti @taoliu032
