### 2024-02-25 moritzobenauer/ProjectRaccoon#10 @taoliu032
I got the same error and it was also resolved by installing from the directory.
