### 2024-02-26 moritzobenauer/ProjectRaccoon#12 @taoliu032
I have similar comments regarding the documentation and automated tests for the JOSS review process, so I decided I will list them here rather than opening a new issue:

1. The current documentation in the README file contains a lot of information, which can perhaps be a bit tedious for first-time users to get through. So I recommend creating a "Table of Content" at the top to clearly demonstrate the content and hierarchy of the documentation -- it will make it much more efficient and easier to navigate it. For example, I was wondering how to specify the parameters for "self-avoiding random walk" and eventually found some instructions at the middle of the README file. It would have been much easier if there were a table of content :)
2. The software is equipped with many useful functions and a very convenient command line user interface (I really like the interface, bravo!). I would recommend putting in some more effort to create more (automated) tests to demonstrate these functions. Creating a seperate folder for the tests and referring to the folder when explaining the software functions in the documentation can give users a clearer idea of the structure of the codebase. These can really encourage users to proceed and test the software.

- To be more specific, you have mentioned in the review channel that one could run unit tests by
> ```python
> python -m unittest
> ```
When I did it, it simply spit out
> Ran 0 tests in 0.000s
>
> OK

Passing unit tests is a good thing but this test did not show me clearly what the software is capable of and how to use the functions (if explanation of some function is too lengthy for README, you can consider putting the explanation somewhere else and redirect us to there). This can be resolved by demonstrating more tests :)

3. Some images in the documentation can probably be made smaller, such as the one for the generated polymer peptide conjugate in the Examples section. It was a bit confusing to me as I thought it was a very important result and I got a totally different geometry. If something does not carry very important information, avoid presenting it in a eye-catching way. Plus, having large size images in the documentation makes it a bit more inconvenient to scroll it through.
