### 2024-02-26 #6293 @taoliu032
I had some comments on the documentation and testing as well, and have added to the already opened [issue](https://github.com/moritzobenauer/ProjectRaccoon/issues/12) above. Hope they are helpful!
