### 2024-02-26 moritzobenauer/ProjectRaccoon#12 @kainszs
Good morning and thank you for your input. @taoliu032
Regarding your question about the tests: The tests are located under ```project_raccoon/tests```. In total there are 38 of them, which result in a code coverage of 86%. How and where do you run the tests?

I have just recloned the repository, installed the environment and run the tests in the root directory using

```bash
python -m unittest
````
The result:

![tests](https://github.com/moritzobenauer/ProjectRaccoon/assets/100122494/225236c5-ce01-421a-83b6-a5b493fc0df2)

Unfortunately, I cannot suppress the DeprecationWarning because it comes from a submodule.

Should I write to the tests what is being tested. E.G.: Test data structure, Test reading the sequence file, Test generation of the polymer, $\dots$ ?
