### 2024-02-26 936d648
Strengthen the parsing of the sequence file
Following [this discussion](https://github.com/moritzobenauer/ProjectRaccoon/issues/11), linked to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6293), this is a commit showing an example of how to strengthen the parsing of the sequence file by raising exceptions containing why and where the parsing fails.
