### 2024-02-26 moritzobenauer/ProjectRaccoon#11 @lorenzo-rovigatti
I have made a PR to show you a quick&dirty way of doing it. Feel free to either accept it or use it as inspiration. The idea is to generate meaningful errors containing information that the user can leverage to fix their sequence file. It can be done in a cleaner way, but I think this should suffice in your case.
