### 2024-02-26 moritzobenauer/ProjectRaccoon#12 @lorenzo-rovigatti
Yeah, I think that letting the user know what's being tested would help.
