### 2024-02-26 #6293 @lorenzo-rovigatti
> Hi, we have noticed that with increasingly larger sequences, we get performance issues when generating the polymer.
>
> We are testing the module against the example sequence. We have noticed that the actions sometimes abort due to a runtime error, i.e. that the tests exceed the 6h limit. This is the case with almost every fourth test.
>
> You can run the tests via
>
> ```shell
> python -m unittest
> ```
>
> I attribute this to the fact that the generation of the sequence by the self-avoiding-random walk is in a continuous loop because no displacement vector is found that fulfills the minimum distance condition.
>
> The polymer is generated in `project_raccoon/src/functions/standard.py` using the functions '`generate_file()` and `get_semi_random_walk_shift()`. We store all coordinates of the polymer in a matrix and generate a vector k→ so that the coordinates of the new monomer buildingblock added with k→ exceed at least the threshold as distance. Do you have any idea how we can fix this without increasing the threshold? We would be grateful for any ideas.
>
> ![aborted_action](https://private-user-images.githubusercontent.com/100122494/307511777-c83a54c7-aaf6-47d0-ae7d-363bf7608653.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MDg5NTA4MjEsIm5iZiI6MTcwODk1MDUyMSwicGF0aCI6Ii8xMDAxMjI0OTQvMzA3NTExNzc3LWM4M2E1NGM3LWFhZjYtNDdkMC1hZTdkLTM2M2JmNzYwODY1My5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjQwMjI2JTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI0MDIyNlQxMjI4NDFaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT1kMDY5YjU5ZDVhNTRiMWViYTk0Yzk5MmNiMzJjMzM1ZGU3YTJiMmRjNmFhMDc2M2IzZjA2OWU2NTdmNTU3NDI5JlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCZhY3Rvcl9pZD0wJmtleV9pZD0wJnJlcG9faWQ9MCJ9.Pn9ZcltGAQBe_YpZTFfusfeCbLvu0RKeKtNstmp0yg8)
>
> @lorenzo-rovigatti @taoliu032

I'm not familiar with the inner workings of the code, but I have written codes to generate my fair share of polymer configurations before, and what is perhaps happening here is that sometimes the code gets "trapped" in a configuration where the $k+1$-th monomer cannot be possibly added for geometrical reasons. As a result, the code gets stuck and runs forever. If I were you, I'd add a check that makes the code re-generate the chain from scratch if a monomer is added $N$ times without success. I'm not sure what $N$ should be, but something like $10^3$, $10^4$ looks like a sensible choice.
