### 2024-02-26 moritzobenauer/ProjectRaccoon#11 @kainszs
Thank you very much for your contribution. I think we'll take it for now and as soon as we have more time, we'll look at it again.  Because in the long term we want to extend the functionality of the software.
