### 2024-02-27 ba828de
Add a check that makes the code re-generate the polymer chain from scratch if a monomer is added 5e3  times without success. Now the sequence can be generated 100 times without errors.
