+++ project_raccoon/src/functions/standard.py
@@ -35,7 +35,9 @@ def generate_sequence(monomers: Monomers, fpath: str) -> Sequence:
                 spl = line.split(":")
 
                 if len(spl) != 4:
-                    raise Exception(f"Sequence line {i}: expected 4 colon-separated values, found {len(spl)}")
+                    raise Exception(
+                        f"Sequence line {i}: expected 4 colon-separated values, found {len(spl)}"
+                    )
 
                 res, resolution, inv, rep = spl
 
@@ -46,7 +48,9 @@ def generate_sequence(monomers: Monomers, fpath: str) -> Sequence:
                 elif resolution == "CG":
                     resolution_lookup = "coarse_grained"
                 else:
-                    raise Exception(f"Unsupported resolution {resolution}; only AA, UA and CG are currently supported")
+                    raise Exception(
+                        f"Unsupported resolution {resolution}; only AA, UA and CG are currently supported"
+                    )
 
                 index.append(
                     monomers.index({"name": res, "resolution": resolution_lookup})
@@ -55,18 +59,23 @@ def generate_sequence(monomers: Monomers, fpath: str) -> Sequence:
                 try:
                     inverted.append(bool(int(inv)))
                 except ValueError:
-                    raise Exception(f"Sequence line {i}: cannot convert {inv} (Inverted field) to a boolean value")
-                
+                    raise Exception(
+                        f"Sequence line {i}: cannot convert {inv} (Inverted field) to a boolean value"
+                    )
+
                 try:
                     rep_int = int(rep)
                 except ValueError:
-                    raise Exception(f"Sequence line {i}: cannot convert {rep} (Repeats field) to an integer value")
-                
+                    raise Exception(
+                        f"Sequence line {i}: cannot convert {rep} (Repeats field) to an integer value"
+                    )
+
                 if rep_int < 0:
-                    raise Exception(f"Sequence line {i}: the Repeats field should contain a positive integer")
+                    raise Exception(
+                        f"Sequence line {i}: the Repeats field should contain a positive integer"
+                    )
 
                 reps.append(rep_int)
-
 
     return Sequence(index, inverted, reps)
 
@@ -127,26 +136,29 @@ def get_semi_random_walk_shift(
     trr: float,
     cshift: List[float],
     shift_conf: List[float],
-):
+    max_iter: int = 1e3,
+) -> np.ndarray:
     """Self-Avoiding Random Walk to prevent infinite forces upon energy minimization.
-         Combines RandShift() and MinimalDistance(). Returns k, which can be used to add onto the new monomer.
-            Treshshold of trr=1 should be sufficient to prevent infinite forces.
+        Combines RandShift() and MinimalDistance(). Returns k, which can be used to add onto the new monomer.
+        Treshshold of trr=1 should be sufficient to prevent infinite forces. If the function fails to find a valid shift, it will return None.
 
     Args:
         - `polypetide_coordinates (np.ndarray, (3,N))`: Array of all coordinates of the atoms in the previous monomers.
         - `monomer (Monomer)`: Monomer from raccoon.src.data
         - `trr (float)`: Threshold for minimal distance.
         - `cshift (List(float))`: Cartesian shift.
-        - `shift_conf (List(float))`: Shift configuration for RandShift().
+        - `shift_conf (List(float))`: Shift configuration for RandShift()
+        - `max_iter (int)`: Maximum number of iterations.
 
     Returns:
-        - `k (np.ndarray)`: 3d vector with shape (3,)
+        - `k (np.ndarray)`: 3d vector with shape (3,) if successful, else False.
     """
 
     monomer_coordinates = monomer.coordinates_to_numpy() + cshift
[…]
