### 2024-02-27 moritzobenauer/ProjectRaccoon#12 @kainszs
Would it be sufficient if we wrote in the documentation that you can execute the unittests using

```bash
python -m unittest -v
```
 to display what is tested with the docstrings of the methods?
