### 2024-02-27 2db726f
extend docstrings of the test functions
