+++ project_raccoon/tests/__main__.py
@@ -7,5 +7,5 @@ if __name__ == "__main__":
 
     loader = unittest.TestLoader()
     testSuite = loader.discover(start_dir=root_dir, pattern="test_*.py")
-    runner = unittest.TextTestRunner()
+    runner = unittest.TextTestRunner(verbosity=2)
     runner.run(testSuite)
