### 2024-02-27 b5db845
Remove unnecessary print statements in standard.py
