### 2024-02-27 moritzobenauer/ProjectRaccoon#12 @taoliu032
> Regarding your question about the tests: The tests are located under project_raccoon/tests. In total there are 38 of them, which result in a code coverage of 86%. How and where do you run the tests?

I just ran the test by  `python -m unittest` as mentioned above, and your suggestion `python -m unittest -v` did print out more details of the tests. That was helpful @kainszs
