### 2024-02-29 moritzobenauer/ProjectRaccoon#12 @lorenzo-rovigatti
I did the same and the output looks fine.
