### 2024-02-29 moritzobenauer/ProjectRaccoon#12 @moritzobenauer
Thank you for the seamless reviewing process so far. I worked on the README-related issues:

@lorenzo-rovigatti
1) Added a paragraph further describing the addition of monomers to the README file. All information previously stated in the monomers.dat file is now easily accessible in the README file.
2) Added a section that describes how to run the unit test to the README file.
3) Added a section describing the _GROMACS_ testing and the contents of the gmx folder. Contents of the example folder are further described in the corresponding sections of the README file.

@taoliu032
1) Added ToC to the README file.
2) Resolved by @kainszs
3) Adjusted the image sizes in the README files. Added clarification that the resulting geometry can be different from what is shown in the example picture.

I added these changes to a second branch called **Moritz0229**. @kainszs opened a pull request if you are happy with the changes.

Please don't hesitate to let us know if you have further comments.
