### 2024-02-29 #6293 @kainszs
I have also tested the installation using:

```bash
python setup.py sdist
pip install dist/project_raccoon-1.0.1.tar.gz
```

As soon as we have the new release out, the package should work properly, even if it is installed via pip.

Should we wait for the DOI? @moritzobenauer
