### 2024-03-04 #6293 @moritzobenauer
Dear @jromanowska ,

the new release of Project RACCOON with the tag  **v1.0.2** is released.
I archived this version to [Zenodo](https://zenodo.org/records/10776652) with the following DOI:

https://doi.org/10.5281/zenodo.10776652
