### 2024-03-06 #6293 @jromanowska
@moritzobenauer - I've noticed that one of the references lacks a DOI (https://onlinelibrary.wiley.com/doi/10.1002/0471721204.ch8) - could you fix that and re-create the PDF?
