+++ JOSS/paper.bib
@@ -101,7 +101,8 @@ year = {2021}
   volume={44},
   pages={161--179},
   year={2003},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/0471721204.ch8}
 }
 
 @article{avogadro,
