+++ JOSS/paper.bib
@@ -123,5 +123,6 @@ year = {2021}
  year = {2009},
  isbn = {1441412697},
  publisher = {CreateSpace},
- address = {Scotts Valley, CA}
+ address = {Scotts Valley, CA},
+ doi={10.5555/1593511}
 }
