### 2024-03-10 #6293 @Kevin-Mattheus-Moerman
@moritzobenauer I have checked your paper, this archive, your repository, and the archive link. Most seems in order. I only have the below points that require your attention:

- [ ] Please edit the author order for the archive to match the paper author order.
- [ ] Please edit the archive listed version tag to match `v1.0.2` it currently says `1.0.2`.
- [ ] Looking at your repository I do not see a clear link to any contributing guidelines. Would you be able to add a simple CONTRIBUTING.md file and link to it from the README.md? This website features some templates: https://contributing.md/example/
