+++ contributing/contributing.md
@@ -24,10 +24,9 @@ All types of contributions are encouraged and valued. See the [Table of Contents
 ## Code of Conduct
 
 This project and everyone participating in it is governed by the
-[Project RACCOON Code of Conduct](https://github.com/moritzobenauer/ProjectRaccoonblob/master/CODE_OF_CONDUCT.md).
+[Project RACCOON Code of Conduct](https://github.com/moritzobenauer/ProjectRaccoonblob/contributing/codeofconduct.md).
 By participating, you are expected to uphold this code. Please report unacceptable behavior
 to <>.
-
 
 ## I Have a Question
 
