+++ contributing/contributing.md
@@ -24,7 +24,7 @@ All types of contributions are encouraged and valued. See the [Table of Contents
 ## Code of Conduct
 
 This project and everyone participating in it is governed by the
-[Project RACCOON Code of Conduct](https://github.com/moritzobenauer/ProjectRaccoon/contributing/codeofconduct.md).
+[Project RACCOON Code of Conduct](https://github.com/moritzobenauer/ProjectRaccoon/blob/main/contributing/codeofconduct.md).
 By participating, you are expected to uphold this code.
 
 ## I Have a Question
