### 2024-03-10 #6293 @moritzobenauer
Hello @Kevin-Mattheus-Moerman ,

I worked on your suggestions:

>
> * [ x]  Please edit the author order for the archive to match the paper author order.

Order is now MLO, KNS, PB, and FS matching the paper.

> * [ x]  Please edit the archive listed version tag to match `v1.0.2` it currently says `1.0.2`.

Version tag is set to v1.0.2.

> * [x ]  Looking at your repository I do not see a clear link to any contributing guidelines. Would you be able to add a simple CONTRIBUTING.md file and link to it from the README.md? This website features some templates: [contributing.md/example](https://contributing.md/example/)

Added contributing.md and codeofconduct.md to the project and linked created a link in the README.md under 'Supported Building Blocks' where we explain contributions to the project.
